Move blocking multipart response logic to `NettySyncToResponseBody`
